+++ cognite/neat/_rules/models/dms/_rules_input.py
@@ -193,6 +193,10 @@ class DMSInputView(InputComponent[DMSView]):
     neatId: str | URIRef | None = None
     logical: str | URIRef | None = None
 
+    def __post_init__(self):
+        if self.in_model is None:
+            self.in_model = True
+
     @classmethod
     def _get_verified_cls(cls) -> type[DMSView]:
         return DMSView